+++ CHANGES.md
@@ -1,1 +1,3 @@
-Updated imports and exports
+- Added sketch point for edit only mode.
+- Fixed start editing from between line segments.
+- Fixed autofocusing.